requirements.txt
django==3.2
djangorestframework==3.13.1
django-debug-toolbar==3.2.1
django-phonenumber-field==6.2.0
Pillow==8.2.0
environs[django]==9.3.2
phonenumbers==8.12.50
